% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\name{add.anomaly}
\alias{add.anomaly}
\title{Add anomaly by county}
\usage{
add.anomaly(in.data, vars, analysis.counties)
}
\arguments{
\item{in.data}{The data set to add anomaly data for. Must contain a 
location field that corresponds to the analysis.counties input.
Only locations in the analysis.counties will have calculations performed}

\item{vars}{The variables in the data set that need the anomaly calculations}

\item{analysis.counties}{The locations for which to calculate anomalies}
}
\value{
in.data
}
\description{
Calculate deviation from the dataset-wide mean value for each location
}
