% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{conus_weather}
\alias{conus_weather}
\title{CONUS mean temp and precipitation by county and month}
\format{
A dataframe with 4699296 observations of 9 variables
\describe{
  \item{name}{character name of the state}
  \item{fips}{character 5 digit FIPS code for the county}
  \item{state_fips}{character two digit FIPS code for the state}
  \item{cofips}{The three-digit fips county code.(Examples: 099, 005)}
  \item{state_noaa}{character two digit state code (NOAA code NOT FIPS)}
  \item{year}{The year (four-digits).(Examples: 2002, 2014)}
  \item{month}{Three letter abbreviation for month of observation, capitalized}
  \item{mean_temp}{Monthly mean temperature in degrees Fahrenheit.}
  \item{precip}{Monthly total precipitation in inches?}
}
}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/cirs/climdiv/}
}
\usage{
conus_weather
}
\description{
Period of record is 1895 through latest month available, updated monthly.
}
\details{
County values in nClimDiv were derived from area-weighted averages of
grid-point estimates interpolated from station data. The Global Historical
Climatology Network (GHCN)  Daily dataset is the source
of station data for nClimDiv.

Two counties changed names or merged over the period 2000-present. To allow
this data to be joined with the CDC WNV data the fips fields are edited
to match, so have fips codes "46102/46113" and "51019/51515". The District
of Columbia is not uniquely represented in the NOAA datasets. In the raw data
DC is fips 24511 (NOAA code 18511). This is changed to fips 11001 to match
CDC data.
}
\keyword{datasets}
