% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldas.april.R
\name{monthly.update}
\alias{monthly.update}
\title{Monthly Update}
\usage{
monthly.update(nldas.path, year, month, month.label)
}
\arguments{
\item{nldas.path}{The path to the NLDAS monthly data}

\item{year}{The year of the update}

\item{month}{The month of the update}

\item{month.label}{The label to apply to the updated data}
}
\description{
Monthly Update
}
