% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{nldas.centroid.lookup}
\alias{nldas.centroid.lookup}
\title{NLDAS County Centroid Lookup}
\format{An object of class \code{data.frame} with 3108 rows and 7 columns.}
\source{
tl_2017_us_county_LOWER_48.shp, NLDAS Soil moisture data
}
\usage{
nldas.centroid.lookup
}
\description{
Create a file that can be used to identify which NLDAS cells
correspond to the county centroid for every county in the US.
Census data used the US Census 2017 shapefile. Centroids
were calculated using ArcGIS 10.6 using the Feature to Point (Data Management) tool.
The latitude and longitude for each NLDAS cell were exported to a point file with
information on the row and column from the NLDAS grid.
Sixty-eight centroids that fell outside the NLDAS grid were moved to a 
nearby appropriate location within the grid. In part, some county boundaries
extend into bodies of water, so the centroid was in the water, rather than in
the actual county. In the case of some, they were on islands too small to 
fill a grid cell (e.g., Florida-Dukes (Keys) moved to the mainland,
and Massachusetts-Nantucket was moved to Martha's Vineyard.
A 'moved' column indicates counties with moved centroids.
These were then merged with the census centroid file using the
spatial join (analysis tool), and exported to .csv format.
}
\keyword{datasets}
