% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{nldas.may}
\alias{nldas.may}
\title{NLDAS May Soil Moisture Data}
\format{An object of class \code{data.frame} with 68376 rows and 6 columns.}
\source{
\url{https://disc.gsfc.nasa.gov/}
}
\usage{
nldas.may
}
\description{
See NLDAS April documentation, generated with the monthly.update function in nldas.april.R file
}
\keyword{datasets}
