% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{nldas.SOILM}
\alias{nldas.SOILM}
\title{NLDAS Soil Moisture Data}
\format{An object of class \code{data.frame} with 798756 rows and 6 columns.}
\source{
\url{https://disc.gsfc.nasa.gov/}
}
\usage{
nldas.SOILM
}
\description{
Soil moisture data were downloaded from NLDAS NOAH and read into R. Only data
from the county centroid were used. However this may not be representative of
the soil moisture in each county, and a county average may perform better
(empirically there was no advantage to using a county average over a county
centroid in Keyel et al. 2019)
}
\details{
See NLDAS April Soil Moisture Data for details \code{\link{nldas.april}}
This data set differs from that one in providing data for all months from
Jan 1999 - April 2020.
}
\keyword{datasets}
