% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.quarterly.R
\name{process.us.quarterly}
\alias{process.us.quarterly}
\title{NOTE: The processing did not handle leap years appropriately,
and therefore the 1st quarter includes data through March 31 in leap years
and April 1 in non-leap years due to a bug in the code.
This happened because the day-of-year breaks were initialized on 2020,
without thought that it was being applied to non-leap years as well.
The approach used here has the benefit of matching day-of-year in the resulting aggregates.
However, consistency in the time periods was desired, so this approach has been corrected.}
\usage{
process.us.quarterly()
}
\description{
NOTE: The processing did not handle leap years appropriately,
and therefore the 1st quarter includes data through March 31 in leap years
and April 1 in non-leap years due to a bug in the code.
This happened because the day-of-year breaks were initialized on 2020,
without thought that it was being applied to non-leap years as well.
The approach used here has the benefit of matching day-of-year in the resulting aggregates.
However, consistency in the time periods was desired, so this approach has been corrected.
}
