% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{spi1}
\alias{spi1}
\title{One month SPI or SPEI}
\format{
A dataframe with 913164 observations of 7 variables
\describe{
  \item{polygon}{Identifier from WWDT site}
  \item{state}{character name of the state}
  \item{county} {character names of the county}
  \item{date}{Date middle of the month}
  \item{year}{The year (four-digits).(Examples: 2002, 2014)}
  \item{month}{double numeric value of month}
  \item{spi1}{double standardized precipitation index}
  \item{spei1}{double standardized precipitation evaporation index}
}
}
\source{
\url{https://wrcc.dri.edu/wwdt/data/REGION/}
}
\usage{
spi1
}
\description{
Period of record is 1997 through latest month available, updated monthly.
}
\keyword{datasets}
