% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/april.gridmet.R, R/gridmet_updater.R
\name{us.month.update}
\alias{us.month.update}
\title{Update with a new month}
\usage{
us.month.update(STUFF)

us.month.update(STUFF)
}
\description{
Update with a new month

Update with a new month
}
