% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{census.data}
\alias{census.data}
\title{Census Data by year}
\source{
#**# FILL IN
}
\usage{
census.data()
}
\description{
Census data #**# FILL IN DOCUMENTATION
See census.data.R for details on processing from data set received from Drew Tyre
}
