% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{fips.lookup}
\alias{fips.lookup}
\title{FIPS Lookup Table}
\format{An object of class \code{data.frame} with 3108 rows and 4 columns.}
\source{
\url{https://www.census.gov/geographies/reference-files/2017/demo/popest/2017-fips.html}
}
\usage{
fips.lookup
}
\description{
FIPS codes and CDC locations to facilitate merging data sets by either FIPS or by location
FIPS Lookup was derived from:
https://www.census.gov/geographies/reference-files/2017/demo/popest/2017-fips.html
Link: 2017 State, County, Minor Civil Division, and Incorporated Place FIPS Codes
file: all-geocodes-v2017.xlsx
Changes were made manually in Excel in order to condense to four columns ready to use for
the CDC forecast challenge.
"South Dakota-Oglala Lakota/Shannon" was replaced with "South Dakota-Oglala Lakota" using R
county gives the state and county, fips the census fips (without a leading zero), the state
fips (also without a leading zero), and the location (which is the same as the county field,
but with the name location for easier joining to forecast challenge data)
}
\keyword{datasets}
