% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{us.quarterly}
\alias{us.quarterly}
\title{US Quarterly GRIDMET data}
\format{An object of class \code{data.frame} with 68376 rows and 27 columns.}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/IDAHO_EPSCOR_GRIDMET}
}
\usage{
us.quarterly
}
\description{
Data from the GRIDMET project, downloaded from Google Earth Engine daily by county using the
GRIDMET Viewer and Downloader Version 1.1 tool in the ArboMAP package (www.github.com/ecograph/ArboMAP).
The downloader was modified to include the COUNTYFP field in the output for the 3 counties where it mattered:
L199 var oldnames = ["NAME", "COUNTYFP", "doy", "year", "tminc", "tmeanc", "tmaxc", "pr", "rmean", "vpd"];
L200 var newnames = ["district", "COUNTYFP", "doy", "year", "tminc", "tmeanc", "tmaxc", "pr", "rmean", "vpd"];
COUNTYFP was then used to disambiguate the 6 county pairs that had the same names for the county and the city.
Converted to quarterly data using the RF1 package tool convert.env.data tool designed to take data
from the ArboMAP daily format and put it into the RF1 input format.
The .csv downloaded from Google Earth Engine is included in the data-raw folder for New York State (New York36.csv)
The converted .rda file is also included New York36.rda.
The .csv and .rda files for the other states can be made available upon request (as of 2020-05-19).
(not included due to their large file sizes to keep the repository small size-wise)
Currently does not include the anomaly data (these were calculated directly in R)
}
\details{
_1 refers to first quarter data, from Jan - Mar, _APRIL refers to data from April

See april.gridmet.R for processing details.
}
\keyword{datasets}
