% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.quarterly.R, R/wnv_data_docs.R
\docType{data}
\name{us.quarterly}
\alias{us.quarterly}
\title{NOTE: The processing did not handle leap years appropriately,
and therefore the 1st quarter includes data through March 31 in leap years
and April 1 in non-leap years due to a bug in the code.
This happened because the day-of-year breaks were initialized on 2020,
without thought that it was being applied to non-leap years as well.
The approach used here has the benefit of matching day-of-year in the resulting aggregates.
However, consistency in the time periods was desired, so this approach has been corrected.}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/IDAHO_EPSCOR_GRIDMET}
}
\usage{
us.quarterly()

us.quarterly()
}
\description{
Data from the GRIDMET project, downloaded from Google Earth Engine daily by county using the
GRIDMET Viewer and Downloader Version 1.1 tool in the ArboMAP package (www.github.com/ecograph/ArboMAP).
The downloader was modified to include the COUNTYFP field in the output for the 3 counties where it mattered:
L199 var oldnames = ["NAME", "COUNTYFP", "doy", "year", "tminc", "tmeanc", "tmaxc", "pr", "rmean", "vpd"];
L200 var newnames = ["district", "COUNTYFP", "doy", "year", "tminc", "tmeanc", "tmaxc", "pr", "rmean", "vpd"];
COUNTYFP was then used to disambiguate the 6 county pairs that had the same names for the county and the city.
Converted to quarterly data using the RF1 package tool convert.env.data tool designed to take data
from the ArboMAP daily format and put it into the RF1 input format.
The .csv downloaded from Google Earth Engine is included in the data-raw folder for New York State (New York36.csv)
The converted .rda file is also included New York36.rda.
The .csv and .rda files for the other states can be made available upon request (as of 2020-05-19).
(not included due to their large file sizes to keep the repository small size-wise)
Currently does not include the anomaly data (these were calculated directly in R)
}
\details{
_1 refers to first quarter data, from Jan - Mar, _APRIL refers to data from April

See april.gridmet.R for processing details.
}
